Semana 14

+++ Proyecto_reservas/app.py
@@ -7,13 +7,16 @@ from reservas.gestor_clientes import GestorClientes
 from reservas.gestor_reservas import GestorReservas
 from reservas.gestor_mesas import GestorMesas
 from reservas.gestor_horarios import GestorHorarios
-from flask_sqlalchemy import SQLAlchemy
+#from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager, login_user, logout_user, login_required, current_user
+from reservas.models import Usuario
 from reservas.gestor_persistencia import (
     guardar_txt, leer_txt,
     guardar_json, leer_json,
     guardar_csv, leer_csv
 )
 from conexion.conexion import obtener_conexion
+from werkzeug.security import generate_password_hash, check_password_hash
 
 
 # CONFIGURACIÓN DE LA APLICACIÓN
@@ -22,12 +25,21 @@ app = Flask(__name__)
 
 # Clave secreta para formularios Flask-WTF
 app.config['SECRET_KEY'] = 'mi_clave_secreta'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///reserva.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db = SQLAlchemy (app)
+#app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///reserva.db'
+#app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+#db = SQLAlchemy (app)
 
 # Inicializar base de datos
 init_db()
+
+
+#Inicializa Flask-Login
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = "login"
+
+login_manager.login_message = "⚠️ Debes iniciar sesión para acceder a esta página"
+login_manager.login_message_category = "warning"
 
 # Crear gestores
 gestor_clientes = GestorClientes()
@@ -39,7 +51,7 @@ gestor_horarios = GestorHorarios()
 # PÁGINAS PRINCIPALES
 
 # Página de inicio
-@app.route('/')
+@app.route('/index')
 def inicio():
     return render_template("index.html")
 
@@ -66,7 +78,12 @@ def calendario():
 
 # Panel administrativo
 @app.route('/admin')
+@login_required
 def admin():
+
+    if current_user.rol != "admin":
+        flash("Acceso no autorizado", "danger")
+        return redirect(url_for("login"))
 
     total_reservas = len(gestor_reservas.listar_reservas())
     total_clientes = len(gestor_clientes.listar_clientes())
@@ -86,6 +103,7 @@ def menu():
 
 # Listar clientes
 @app.route("/clientes")
+@login_required
 def clientes_listar():
 
     clientes = gestor_clientes.listar_clientes()
@@ -98,6 +116,7 @@ def clientes_listar():
 
 # Crear cliente
 @app.route("/clientes/nuevo", methods=["GET", "POST"])
+@login_required
 def cliente_nuevo():
 
     form = ClienteForm()
@@ -131,6 +150,7 @@ def cliente_nuevo():
 
 # Editar cliente
 @app.route("/clientes/editar/<int:id>", methods=["GET", "POST"])
+@login_required
 def editar_cliente(id):
 
     cliente = gestor_clientes.obtener_cliente_por_id(id)
@@ -169,6 +189,7 @@ def editar_cliente(id):
 
 # Eliminar cliente
 @app.route("/clientes/eliminar/<int:id>")
+@login_required
 def eliminar_cliente(id):
 
     gestor_clientes.eliminar_cliente(id)
@@ -180,26 +201,35 @@ def eliminar_cliente(id):
 
 # Buscar cliente
 @app.route("/clientes/buscar", methods=["GET", "POST"])
+@login_required
 def buscar_cliente():
 
     resultados = []
+    mensaje = None
 
     if request.method == "POST":
 
-        texto = request.form["texto"]
+        texto = request.form.get("texto")
 
-        resultados = gestor_clientes.buscar_cliente(texto)
+        if not texto or texto.strip() == "":
+            mensaje = "⚠️ Debe ingresar un nombre para buscar"
+        else:
+            resultados = gestor_clientes.buscar_cliente(texto)
+
+            if not resultados:
+                mensaje = "❌ No se encontraron clientes"
 
     return render_template(
         "buscar_cliente.html",
-        resultados=resultados
+        resultados=resultados,
+        mensaje=mensaje
     )
-
 
 # CRUD RESERVAS
 
 # Listar reservas
 @app.route("/reservas")
+@login_required
 def reservas():
 
     reservas = gestor_reservas.listar_reservas()
@@ -212,6 +242,7 @@ def reservas():
 
 # Crear reserva
 @app.route("/reservas/nuevo", methods=["GET", "POST"])
+@login_required
 def reserva_nuevo():
 
     form = ReservaForm()
@@ -251,8 +282,35 @@ def reserva_nuevo():
 
     return render_template("reserva_form.html", form=form)
 
+# Buscar reserva
+@app.route("/reservas/buscar", methods=["GET", "POST"])
+@login_required
+def buscar_reserva():
+
+    resultados = []
+    mensaje = None
+
+    if request.method == "POST":
+
+        texto = request.form.get("texto")
+
+        if not texto or texto.strip() == "":
+            mensaje = "⚠️ Debe ingresar un nombre para buscar"
+        else:
+            resultados = gestor_reservas.buscar_reserva(texto)
+
+            if not resultados:
+                mensaje = "❌ No se encontraron reservas"
+
+    return render_template(
+        "buscar_reserva.html",
+        resultados=resultados,
+        mensaje=mensaje
+    )
+
 # Editar reserva
 @app.route("/reservas/editar/<int:id>", methods=["GET","POST"])
+@login_required
 def editar_reserva(id):
 
     reserva = gestor_reservas.obtener_reserva(id)
@@ -287,6 +345,7 @@ def editar_reserva(id):
 
 # Eliminar reserva
 @app.route("/reservas/eliminar/<int:id>")
+@login_required
 def eliminar_reserva(id):
 
         gestor_reservas.eliminar_reserva(id)
@@ -300,6 +359,7 @@ def eliminar_reserva(id):
 # MESAS
 
 @app.route("/mesas")
+@login_required
 def mesas():
 
     mesas = gestor_mesas.listar_mesas()
@@ -311,6 +371,7 @@ def mesas():
 
 # Crear nueva mesa
 @app.route("/mesas/nuevo", methods=["GET", "POST"])
+@login_required
 def mesa_nueva():
 
     if request.method == "POST":
@@ -335,6 +396,7 @@ def mesa_nueva():
 
 # Editar mesa
 @app.route("/mesas/editar/<int:id>", methods=["GET","POST"])
+@login_required
 def editar_mesa(id):
 
     mesa = gestor_mesas.obtener_mesa(id)
@@ -352,32 +414,18 @@ def editar_mesa(id):
 
 # Eliminar mesa
 @app.route("/mesas/eliminar/<int:id>")
+@login_required
 def eliminar_mesa(id):
 
     gestor_mesas.eliminar_mesa(id)
     flash("Mesa eliminada","warning")
     return redirect(url_for("mesas"))
 
-# Buscar reserva
-@app.route("/reservas/buscar", methods=["GET", "POST"])
-def buscar_reserva():
-
-    resultados = []
-
-    if request.method == "POST":
-
-        texto = request.form.get("texto")
-
-        resultados = gestor_reservas.buscar_reserva(texto)
-
-    return render_template(
-        "buscar_reserva.html",
-        resultados=resultados
-    )
 
 # HORARIOS
 
 @app.route("/horarios")
+@login_required
 def horarios():
 
     horarios = gestor_horarios.listar_horarios()
@@ -388,6 +436,7 @@ def horarios():
 
 # Crear nuevo horario
 @app.route("/horarios/nuevo", methods=["GET", "POST"])
+@login_required
 def horario_nuevo():
 
     if request.method == "POST":
@@ -408,6 +457,7 @@ def horario_nuevo():
 
 # Editar horario
 @app.route("/horarios/editar/<int:id>", methods=["GET","POST"])
+@login_required
 def editar_horario(id):
 
     horario = gestor_horarios.obtener_horario(id)
@@ -425,6 +475,7 @@ def editar_horario(id):
 
 # Eliminar horario
 @app.route("/horarios/eliminar/<int:id>")
+@login_required
 def eliminar_horario(id):
 
     gestor_horarios.eliminar_horario(id)
@@ -433,6 +484,7 @@ def eliminar_horario(id):
 
 # Ruta usuarios
 @app.route("/usuarios/nuevo", methods=["GET","POST"])
+@login_required
 def usuario_nuevo():
 
     if request.method == "POST":
@@ -442,14 +494,22 @@ def usuario_nuevo():
         password = request.form.get("password")
 
         conexion = obtener_conexion()
-        cursor = conexion.cursor()
+
+        if not conexion:
+            flash("Error de conexión a MySQL", "danger")
+            return redirect(url_for("login"))
+
+        cursor = conexion.cursor(dictionary=True)
 
         sql = """
         INSERT INTO usuarios (nombre, email, password)
         VALUES (%s, %s, %s)
         """
 
-        cursor.execute(sql,(nombre,email,password))
+        password_hash = generate_password_hash(password)
+
+        cursor.execute(sql,(nombre,email,password_hash))
+
         conexion.commit()
 
         cursor.close()
@@ -463,9 +523,19 @@ def usuario_nuevo():
 
 # Consultar ususarios
 @app.route("/usuarios")
+@login_required
 def usuarios():
+    
+    if current_user.rol != "admin":
+        flash("Acceso no autorizado", "danger")
+        return redirect(url_for("inicio"))
 
-    conexion = obtener_conexion()
+    conexion = obtener_conexion()   # 👈 FALTABA
+
+    if not conexion:
+        flash("Error de conexión", "danger")
+        return redirect(url_for("inicio"))
+
     cursor = conexion.cursor(dictionary=True)
 
     cursor.execute("SELECT * FROM usuarios")
@@ -478,9 +548,15 @@ def usuarios():
 
 # Actualizar
 @app.route("/usuarios/editar/<int:id>", methods=["GET","POST"])
+@login_required
 def editar_usuario(id):
 
     conexion = obtener_conexion()
+
+    if not conexion:
+        flash("Error de conexión a MySQL", "danger")
+        return redirect(url_for("login"))
+
     cursor = conexion.cursor(dictionary=True)
 
     if request.method == "POST":
@@ -495,7 +571,9 @@ def editar_usuario(id):
         WHERE id_usuario=%s
         """
 
-        cursor.execute(sql,(nombre,email,password,id))
+
+        password_hash = generate_password_hash(password)
+        cursor.execute(sql,(nombre,email,password_hash,id))
         conexion.commit()
 
         cursor.close()
@@ -515,10 +593,16 @@ def editar_usuario(id):
 
 # Eliminar 
 @app.route("/usuarios/eliminar/<int:id>")
+@login_required
 def eliminar_usuario(id):
 
     conexion = obtener_conexion()
-    cursor = conexion.cursor()
+
+    if not conexion:
+        flash("Error de conexión a MySQL", "danger")
+        return redirect(url_for("login"))
+
+    cursor = conexion.cursor(dictionary=True)
 
     cursor.execute("DELETE FROM usuarios WHERE id_usuario=%s",(id,))
     conexion.commit()
@@ -529,6 +613,149 @@ def eliminar_usuario(id):
     flash("Usuario eliminado","warning")
 
     return redirect(url_for("usuarios"))
+
+# Cargar usuario desde MYSQL
+@login_manager.user_loader
+def load_user(user_id):
+
+    conexion = obtener_conexion()
+
+    if not conexion:
+        return None   # ✅ SOLO ESTO
+
+    cursor = conexion.cursor(dictionary=True)
+
+    cursor.execute("SELECT * FROM usuarios WHERE id_usuario=%s", (int(user_id),))
+    user = cursor.fetchone()
+
+    cursor.close()
+    conexion.close()
+
+    if user:
+        return Usuario(
+            user["id_usuario"],
+            user["nombre"],
+            user["email"],
+            user["password"],
+            user["rol"]
+
+        )
+
+    return None
+
+# Registro de usuarios
+@app.route("/registro", methods=["GET","POST"])
+def registro():
+
+    if request.method == "POST":
+
+        nombre = request.form.get("nombre")
+        email = request.form.get("email")
+        password = request.form.get("password")
+
+        conexion = obtener_conexion()
+        if not conexion:
+            flash("Error de conexión a la base de datos", "danger")
+            return redirect(url_for("login"))
+
+        cursor = conexion.cursor(dictionary=True)
+
+        if not email or not password:
+            flash("Complete todos los campos", "warning")
+            return redirect(url_for("login"))
+        
+        # 🔍 Verificar si el correo ya existe
+        cursor.execute("SELECT * FROM usuarios WHERE email=%s", (email,))
+        existe = cursor.fetchone()
+
+        if existe:
+            cursor.close()
+            conexion.close()
+            flash("El correo ya está registrado", "warning")
+            return redirect(url_for("registro"))
+
+        # Encriptar contraseña
+        password_hash = generate_password_hash(password, method='pbkdf2:sha256', salt_length=16)
+
+        # Insertar usuario
+        rol = "usuario"
+
+        sql = "INSERT INTO usuarios (nombre,email,password,rol) VALUES (%s,%s,%s,%s)"
+        cursor.execute(sql,(nombre,email,password_hash,rol))
+
+        conexion.commit()
+        cursor.close()
+        conexion.close()
+
+        flash("Usuario registrado correctamente","success")
+        return redirect(url_for("login"))
+
+    return render_template("registro.html")
+
+# Login
+@app.route("/", methods=["GET","POST"])
+def login():
+
+    if request.method == "POST":
+
+        email = request.form.get("email")
+        password = request.form.get("password")
+
+        if not email or not password:
+            flash("Complete todos los campos", "warning")
+            return redirect(url_for("login"))
+
+        conexion = obtener_conexion()
+        if not conexion:
+            flash("Error de conexión a la base de datos", "danger")
+            return redirect(url_for("login"))
+
+        cursor = conexion.cursor(dictionary=True)
+
+        cursor.execute(
+            "SELECT * FROM usuarios WHERE email=%s",
+            (email,)
+        )
+
+        user = cursor.fetchone()
+
+        cursor.close()
+        conexion.close()
+
+        # ✅ TODO dentro del POST
+        if user and check_password_hash(user["password"], password):
+
+            usuario = Usuario(
+                user["id_usuario"],
+                user["nombre"],
+                user["email"],
+                user["password"],
+                user["rol"]
+            )
+
+            login_user(usuario)
+            flash("Bienvenido " + usuario.nombre, "success")
+
+            if usuario.rol == "admin":
+                return redirect(url_for("admin"))
+            else:
+                return redirect(url_for("inicio"))
+
+        else:
+            flash("Credenciales incorrectas", "danger")
+            return redirect(url_for("login"))
+
+    return render_template("login.html")
+
+# Logout cerrar sesión
+@app.route("/logout")
+@login_required
+def logout():
+
+    logout_user()
+    flash("Sesión cerrada","info")
+    return redirect(url_for("login"))
+
 
 # PERSISTENCIA ARCHIVOS
 
@@ -573,5 +800,5 @@ def datos():
 
 # EJECUTAR APLICACIÓN
 if __name__ == '__main__':
-
+    app.config['PROPAGATE_EXCEPTIONS'] = True
     app.run(debug=True)
